Feature: correcciones ultimas
